PROBLEM: logical asset name != asset name
When implementing unicode for assets we introduced 'ext.name' which is
the utf-8 asset name and proto field 'name' now contains the asset
(uu)id.  Field 'ext.logical_asset' is filled by user thus it is always
actually ext.name (== the utf-8 unicode name). It has to be, otherwise
user wouldn't be able to reference the asset, id's are generated
internally.  The problem encountered here was that 'ext.logical_asset'
was used as a key in hash that was indexed by 'name' (field of proto).

SOLUTION: Fix it.
When 'ext.logical_asset' string is received, instead of using it to
lookup the hash, iterate over self->all_assets looking for the
logical_asset string in extended attributes of each item being iterated

Signed-off-by: Karol Hrdina <KarolHrdina@Eaton.com>